+++ GoodFunGame/Assets/@Resources/Prefabs/UI/Popup/UI_Popup_IntroMenu.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: aff580af7d6587f46974c53b13dfac12
+guid: 0e96f5e919706a341a86a3fb41ffa833
 PrefabImporter:
   externalObjects: {}
   userData:
