+++ GoodFunGame/Assets/@Resources/Prefabs/UI/Scene/UI_Scene_Intro.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0e96f5e919706a341a86a3fb41ffa833
+guid: 4e1f8a4f9c1544b47bff1ac59bcc7280
 PrefabImporter:
   externalObjects: {}
   userData:
