+++ GoodFunGame/Assets/@Resources/Sprites/JEH_Stage.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: aff580af7d6587f46974c53b13dfac12
-PrefabImporter:
+guid: 445cda93279adc247be5e2410ed87d76
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
