+++ GoodFunGame/Assets/AddressableAssetsData/AssetGroups/Prefabs.asset
@@ -29,138 +29,6 @@ MonoBehaviour:
     m_SerializedLabels:
     - PreLoad
     FlaggedDuringContentUpdateRestriction: 0
-  - m_GUID: 6a3e0bbb0dfc67240bf0dcf29c1a972f
-    m_Address: UI_Scene_Battle_Multi.prefab
-    m_ReadOnly: 0
-    m_SerializedLabels:
-    - PreLoad
-    FlaggedDuringContentUpdateRestriction: 0
-  - m_GUID: 6493adcb5dad3c045a8c516f56d734ec
-    m_Address: UI_Scene_Battle_Solo.prefab
-    m_ReadOnly: 0
-    m_SerializedLabels:
-    - PreLoad
-    FlaggedDuringContentUpdateRestriction: 0
-  - m_GUID: 0e96f5e919706a341a86a3fb41ffa833
-    m_Address: UI_Scene_Intro.prefab
-    m_ReadOnly: 0
-    m_SerializedLabels:
-    - PreLoad
-    FlaggedDuringContentUpdateRestriction: 0
-  - m_GUID: b0d2181bd19d46e4cb1d1dba53bc9490
-    m_Address: UI_Scene_Select.prefab
-    m_ReadOnly: 0
-    m_SerializedLabels:
-    - PreLoad
-    FlaggedDuringContentUpdateRestriction: 0
-  - m_GUID: 5c0b26f003668864cabe45b89088940b
-    m_Address: UI_Scene_Shop.prefab
-    m_ReadOnly: 0
-    m_SerializedLabels:
-    - PreLoad
-    FlaggedDuringContentUpdateRestriction: 0
-  - m_GUID: 3f1b4f0d5533d0547b7b8617f84dbe42
-    m_Address: UI_Popup_Option.prefab
-    m_ReadOnly: 0
-    m_SerializedLabels:
-    - PreLoad
-    FlaggedDuringContentUpdateRestriction: 0
-  - m_GUID: 733739888e767ea43abe6bb4b6606751
-    m_Address: UI_Popup_Pause.prefab
-    m_ReadOnly: 0
-    m_SerializedLabels:
-    - PreLoad
-    FlaggedDuringContentUpdateRestriction: 0
-  - m_GUID: 6479b0b6bf7630b4babe3daf0cf31938
-    m_Address: UI_Popup_Purchase.prefab
-    m_ReadOnly: 0
-    m_SerializedLabels:
-    - PreLoad
-    FlaggedDuringContentUpdateRestriction: 0
-  - m_GUID: d396d5841cf3a3e499b526f5b66e32ac
-    m_Address: UI_Popup_Status.prefab
-    m_ReadOnly: 0
-    m_SerializedLabels:
-    - PreLoad
-    FlaggedDuringContentUpdateRestriction: 0
-  - m_GUID: da2a5c902e40e934f96457229a0fa7b8
-    m_Address: UI_Popup_Tooltip.prefab
-    m_ReadOnly: 0
-    m_SerializedLabels:
-    - PreLoad
-    FlaggedDuringContentUpdateRestriction: 0
-  - m_GUID: ac4f134ba4b71814cb671ea4bbc28467
-    m_Address: BackspaceBtn.prefab
-    m_ReadOnly: 0
-    m_SerializedLabels:
-    - PreLoad
-    FlaggedDuringContentUpdateRestriction: 0
-  - m_GUID: 20f03240d43c6e540a9e333e734a93e8
-    m_Address: Battle_Panel.prefab
-    m_ReadOnly: 0
-    m_SerializedLabels:
-    - PreLoad
-    FlaggedDuringContentUpdateRestriction: 0
-  - m_GUID: 13ff2d15bcc090344bb3b34e681091c0
-    m_Address: Heart.prefab
-    m_ReadOnly: 0
-    m_SerializedLabels:
-    - PreLoad
-    FlaggedDuringContentUpdateRestriction: 0
-  - m_GUID: 0935b8b8e6aaee344a41d35cdf9b7c05
-    m_Address: HeartList.prefab
-    m_ReadOnly: 0
-    m_SerializedLabels:
-    - PreLoad
-    FlaggedDuringContentUpdateRestriction: 0
-  - m_GUID: 5c24412ebbec3c44e86d4c5a2b9e17ef
-    m_Address: KeyBtn.prefab
-    m_ReadOnly: 0
-    m_SerializedLabels:
-    - PreLoad
-    FlaggedDuringContentUpdateRestriction: 0
-  - m_GUID: 0ff5e559c8722644baf73514bdb4c0c1
-    m_Address: KeyList.prefab
-    m_ReadOnly: 0
-    m_SerializedLabels:
-    - PreLoad
-    FlaggedDuringContentUpdateRestriction: 0
-  - m_GUID: 5cc5750871151ae4a919207950d9f68b
-    m_Address: Option_Resolution.prefab
-    m_ReadOnly: 0
-    m_SerializedLabels:
-    - PreLoad
-    FlaggedDuringContentUpdateRestriction: 0
-  - m_GUID: e772ad91b6ec2e140b975cede38fc071
-    m_Address: Option_Volume.prefab
-    m_ReadOnly: 0
-    m_SerializedLabels:
-    - PreLoad
-    FlaggedDuringContentUpdateRestriction: 0
-  - m_GUID: 1bbca7a00f073e84f9a46d85d09a3010
-    m_Address: Select_Button.prefab
-    m_ReadOnly: 0
-    m_SerializedLabels:
-    - PreLoad
-    FlaggedDuringContentUpdateRestriction: 0
-  - m_GUID: 3ff86d7accb2720469dfd3bb56d56a74
-    m_Address: Shop_SkillCard.prefab
-    m_ReadOnly: 0
-    m_SerializedLabels:
-    - PreLoad
-    FlaggedDuringContentUpdateRestriction: 0
-  - m_GUID: 9b36e31e963e02343845ce99122cc53f
-    m_Address: SkillBtn.prefab
-    m_ReadOnly: 0
-    m_SerializedLabels:
-    - PreLoad
-    FlaggedDuringContentUpdateRestriction: 0
-  - m_GUID: 785d938dafaa8194e9270284f2d37256
-    m_Address: SkillList.prefab
-    m_ReadOnly: 0
-    m_SerializedLabels:
-    - PreLoad
-    FlaggedDuringContentUpdateRestriction: 0
   m_ReadOnly: 0
   m_Settings: {fileID: 11400000, guid: cedf8085c2bf98c49a1b20eaf2ae0f07, type: 2}
   m_SchemaSet:
