+++ GoodFunGame/Assets/Scripts/UI/Popup/UI_Popup_BattleMenu.cs
@@ -1,0 +1,15 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class UI_Popup_BattleMenu : UI_Popup
+{
+  
+    void Start()
+    {
+ 
+        
+    }
+
+   
+}
