+++ GoodFunGame/Assets/Scripts/UI/Popup/UI_Popup_BattleMenu.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f08e978d398312f45a1035ac465839a1
+guid: d73b7d017e3aa7d48b8c7b2f5e8d2694
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
