+++ GoodFunGame/Assets/Scripts/UI/Scene/UI_Scene_Battle.cs
@@ -7,12 +7,9 @@ public class UI_Scene_Battle : UI_Scene
  
     void Start()
     {
-        
+        // 싱글인지 경쟁인지 협동인지에 따라 다른 ui 뽑기
     }
 
 
-    void Update()
-    {
-        
-    }
+   
 }
