+++ GoodFunGame/Assets/Scripts/UI/Scene/UI_Scene_Select.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f12d75a6412287542a6e41afd56771e2
+guid: bbf5457b678196b4281b78fef919f37c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
