+++ GoodFunGame/Assets/UITestScene.cs
@@ -1,0 +1,30 @@
+using UnityEngine;
+
+public class UITestScene : MonoBehaviour
+{
+    void Start()
+    {
+        if (Main.Resource.Loaded)
+        {
+            Main.Data.Initialize();
+        }
+        else
+        {
+            Main.Resource.LoadAllAsync<UnityEngine.Object>("PreLoad", (key, count, totalCount) =>
+            {
+                Debug.Log($"[TestScene] Load asset {key} ({count} / {totalCount})");
+                if (count >= totalCount)
+                {
+                    Main.Data.Initialize();
+                    InitializeGame();
+                }
+            });
+        }
+    }
+
+    private void InitializeGame()
+    {
+        Main.UI.ShowSceneUI<UI_Scene_Intro>();
+
+    }
+}
