+++ GoodFunGame/Assets/UITestScene.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f08e978d398312f45a1035ac465839a1
+guid: d0b572064495a314b88f91076f73213d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
